+++ ChapaInAppPurchase/src/main/java/com/yosef/chapainapppurchase/payment_type/BasicPayment.java
@@ -3,10 +3,12 @@ package com.yosef.chapainapppurchase.payment_type;
 
 import com.yosef.chapainapppurchase.ChapaError;
 import com.yosef.chapainapppurchase.PaymentType;
-
+/**
+ * Basic payment type
+ */
 public class BasicPayment extends PaymentType {
 
-    public BasicPayment(float amount) {
+    public BasicPayment(double amount) {
         super(amount);
     }
